Move not found errors out of routes and into controllers

+++ src/controllers/mediaRef.ts
@@ -1,6 +1,7 @@
 import { getRepository } from 'typeorm'
 import { MediaRef } from 'entities'
 import { validateClassOrThrow } from 'errors'
+const createError = require('http-errors')
 
 const relations = [
   'authors', 'categories', 'episode', 'owner', 'podcast'
@@ -20,13 +21,24 @@ const createMediaRef = async (obj) => {
 const deleteMediaRef = async (id) => {
   const repository = getRepository(MediaRef)
   const mediaRef = await repository.findOne({ id })
+
+  if (!mediaRef) {
+    throw new createError.NotFound('MediaRef not found')
+  }
+
   const result = await repository.remove(mediaRef)
   return result
 }
 
 const getMediaRef = (id) => {
   const repository = getRepository(MediaRef)
-  return repository.findOne({ id }, { relations })
+  const mediaRef = repository.findOne({ id }, { relations })
+
+  if (!mediaRef) {
+    throw new createError.NotFound('MediaRef not found')
+  }
+  
+  return mediaRef
 }
 
 const getMediaRefs = (query) => {
@@ -43,6 +55,11 @@ const getMediaRefs = (query) => {
 const updateMediaRef = async (obj) => {
   const repository = getRepository(MediaRef)
   const mediaRef = await repository.findOne({ id: obj.id })
+
+  if (!mediaRef) {
+    throw new createError.NotFound('MediaRef not found')
+  }
+
   const newMediaRef = Object.assign(mediaRef, obj)
 
   await validateClassOrThrow(newMediaRef)
