+++ src/controllers/playlist.ts
@@ -1,6 +1,7 @@
 import { getRepository } from 'typeorm'
 import { Playlist } from 'entities'
 import { validateClassOrThrow } from 'errors'
+const createError = require('http-errors')
 
 const relations = [
   'mediaRefs', 'owner'
@@ -20,13 +21,24 @@ const createPlaylist = async (obj) => {
 const deletePlaylist = async (id) => {
   const repository = getRepository(Playlist)
   const playlist = await repository.findOne({ id })
+
+  if (!playlist) {
+    throw new createError.NotFound('Playlist not found')
+  }
+
   const result = await repository.remove(playlist)
   return result
 }
 
 const getPlaylist = async (id) => {
   const repository = getRepository(Playlist)
-  return repository.findOne({ id }, { relations })
+  const playlist = repository.findOne({ id }, { relations })
+
+  if (!playlist) {
+    throw new createError.NotFound('Playlist not found')
+  }
+
+  return playlist
 }
 
 const getPlaylists = (query) => {
@@ -43,6 +55,11 @@ const getPlaylists = (query) => {
 const updatePlaylist = async (obj) => {
   const repository = getRepository(Playlist)
   const playlist = await repository.findOne({ id: obj.id })
+
+  if (!playlist) {
+    throw new createError.NotFound('Playlist not found')
+  }
+
   const newPlaylist = Object.assign(playlist, obj)
 
   await validateClassOrThrow(newPlaylist)
